Back up

+++ admin/modules/product/process.php
@@ -1,6 +1,6 @@
 <?php
     include("../../../core/init.php");
-    $path = '../../../image/';
+    $path = '../../../uploads/';
     if(isset($_POST['themhang'])){
         $mshh=$_POST['MSHH'];
 	    $tensach=$_POST['TenHH'];
@@ -9,17 +9,16 @@
         $gia=$_POST['Gia'];
         $soluong=$_POST['SoLuongHang'];
 
-        
         $i = 0;
-        $name = array();
+        $name = "";
         
         foreach ($_FILES['hinhanh']['name'] as $file) {
-            $tenhinhanh_tmp=$_FILES['hinhanh']['tmp_name'][$i];
-            $name[] = $file;
-            $path_file = $path . $file;
+            $tenhinhanh_tmp = $_FILES['hinhanh']['tmp_name'][$i];
+            $name = $mshh . $file;
+            $path_file = $path . $mshh .$file;
             move_uploaded_file($tenhinhanh_tmp,$path_file);
-            $i++;          
-            $sql1 = "INSERT INTO hinhhanghoa(MSHH,tenhinh,mahinh) VALUE('".$mshh."','".$file."','".$file."')";
+            $i++;
+            $sql1 = "INSERT INTO hinhhanghoa(MSHH,tenhinh,mahinh) VALUE('".$mshh."','".$name."','".$name."')";
             mysqli_query($db, $sql1);
         }
         $sql = "INSERT INTO hanghoa VALUE('".$mshh."','".$tensach."','" .$quycach. "','".$gia."','".$soluong."','" .$maloai. "')";
@@ -33,7 +32,6 @@
          $result = mysqli_query($db, $sql2);
          while($row = mysqli_fetch_assoc($result)){
              unlink($path.$row['tenhinh']);
-
          }
         $sql = "DELETE FROM hanghoa WHERE (MSHH='".$_GET['id']."')";
         mysqli_query($db, $sql);
@@ -55,20 +53,23 @@
             header('Location: ' . $_SERVER['HTTP_REFERER']);
         }
         else{
-             $sql2 = "SELECT * FROM hinhhanghoa WHERE MSHH = '".$mshh."'";
+            $sql2 = "SELECT * FROM hinhhanghoa WHERE MSHH = '".$mshh."'";
             $result = mysqli_query($db, $sql2);
             while($row = mysqli_fetch_assoc($result)){
-                unlink($path.$row['tenhinh']);
+                if(file_exists($path.$row['tenhinh'])) unlink($path.$row['tenhinh']);
             }
             $sql1 = "DELETE FROM hinhhanghoa WHERE (MSHH='".$mshh."')";
             mysqli_query($db, $sql1);
             $i = 0;
+            $name = "";
+
             foreach ($_FILES['hinhanh']['name'] as $file) {
                 $tenhinhanh_tmp=$_FILES['hinhanh']['tmp_name'][$i];
-                $path_file = $path . $file;
+                $path_file = $path . $mshh . $file;
+                $name = $mshh . $file;
                 move_uploaded_file($tenhinhanh_tmp,$path_file);
                 $i++;
-                $sql3 = "INSERT INTO hinhhanghoa(MSHH,tenhinh,mahinh) VALUE('".$mshh."','".$file."','".$file."')";
+                $sql3 = "INSERT INTO hinhhanghoa(MSHH,tenhinh,mahinh) VALUE('".$mshh."','".$name."','".$name."')";
                 mysqli_query($db, $sql3);
             }
             header('Location:../../index.php?module=product');
